:sparkles: Write to stdout if --output is missing

+++ bin/main.dart
@@ -12,13 +12,13 @@ final ArgParser parser = ArgParser()
     'pubspec-lock',
     abbr: 'i',
     help: 'Input pubscpec.lock file',
-    mandatory: true,
+    mandatory: true
   )
   ..addOption(
     'output',
     abbr: 'o',
     help: 'Output combined_licenses.txt',
-    mandatory: true,
+    defaultsTo: null,
   )
   ..addFlag(
     'verbose',
@@ -43,29 +43,28 @@ void main(List<String> arguments) async {
     }
     final Stream<Package> packages = getPackagesStream(pubspecLock.path);
 
-    final File licensesFile = File(path.join(cwd, args['output']));
-    if (await licensesFile.exists()) {
-      await licensesFile.delete();
-      await licensesFile.create();
+    if (args['output'] != null) {
+      final File licensesFile = File(path.join(cwd, args['output']));
+      if (await licensesFile.exists()) {
+        await licensesFile.delete();
+        await licensesFile.create();
+      }
+
+      final IOSink sink = licensesFile.openWrite();
+
+      packages.listen(
+        (Package package) {
+          sink.writeln(_licenseText(package));
+        },
+        onDone: () => sink.close(),
+      );
+
+      if (verbose) stdout.writeln('Wrote: ${licensesFile.path}');
+    } else {
+      packages.listen((Package package) {
+        stdout.writeln(_licenseText(package));
+      });
     }
-
-    final IOSink sink = licensesFile.openWrite();
-
-    packages.listen(
-      (Package package) {
-        sink.writeln('PACKAGE NAME: ${package.name?.trim()}');
-        sink.writeln('PACKAGE VERSION: ${package.version?.trim()}');
-        sink.writeln('PACKAGE HOMEPAGE: ${package.homepage?.trim()}');
-        sink.writeln('LICENSE:');
-        sink.writeln(package.license?.trim());
-        sink.writeln('');
-        sink.writeln('----------');
-        sink.writeln('');
-      },
-      onDone: () => sink.close(),
-    );
-
-    if (verbose) stdout.writeln('Wrote: ${licensesFile.path}');
   } on FormatException catch (err) {
     exitCode = 2;
     stderr.writeln(err.toString());
@@ -77,3 +76,12 @@ void main(List<String> arguments) async {
     if (verbose) rethrow;
   }
 }
+
+String _licenseText(Package package) => <String>[
+      'PACKAGE NAME: ${package.name?.trim()}',
+      'PACKAGE VERSION: ${package.version?.trim()}',
+      'PACKAGE HOMEPAGE: ${package.homepage?.trim()}',
+      'LICENSE:',
+      package.license?.trim() ?? '',
+      '\n----------\n'
+    ].join('\n');
